Revert "Merge branch 'settings' into main"
This reverts commit a558c9eb6243029c0346a7902904f30ba41d5582, reversing
changes made to a2eb2c092e060d8e12c32ad94169618a155693ef.
