--- MealSwap/Settings/AboutView.swift
@@ -1,32 +1,0 @@
-//
-//  AboutView.swift
-//  MealSwap
-//
-//  Created by Bryan Pineda on 4/20/24.
-//
-
-/**
- 
- Bryan Notes:
- 
-   Important: Should be done ASAP (Easy view to complete)
- 
-        - User can view information about our APP
-        - User can scroll
- **/
-
-import SwiftUI
-
-struct AboutView: View {
-    var body: some View {
-        VStack{
-            Text("About Us")
-        }         
-        .navigationTitle("About Us")
-
-    }
-}
-
-#Preview {
-    AboutView()
-}
